+++ Unit 2/Practices/LINEAR_REGRESSION_EXERCISE/Linear_regression.scala
